feat: improve code structure and formatting
